change in customer master

+++ DataAccess/Domain/Transactions/VendorInvoiceTxn/VendorPaymentInvoiceEntity.cs
@@ -10,17 +10,17 @@ namespace DataAccess.Domain.Transactions.VendorInvoiceTxn
     {
         [Key]
         public int ID { get; set; }
-        
+
         [ForeignKey("VendorInvoiceTxnEntity")]
         public int VendorInvoiceTxnID { get; set; }
         public DateTime? paymentDate { get; set; }
         public int? bankID { get; set; }
         public string? oWRMNo1 { get; set; }
         public string? oWRMNo2 { get; set; }
-        public int? rate { get; set; }
-        public int? quantity { get; set; }
-        public int? bankcharges { get; set; }
-        public int? totalAmountInr { get; set; }
+        public decimal? rate { get; set; } // Changed to decimal?
+        public decimal? quantity { get; set; } // Changed to decimal?
+        public decimal? bankcharges { get; set; } // Changed to decimal?
+        public decimal? totalAmountInr { get; set; } // Changed to decimal?
         public string? paymentCurrency { get; set; }
         public decimal? paymentAmount { get; set; }
         public VendorInvoiceTxnEntity? VendorPaymentEntity { get; set; }
